MyFactorial re factored and comments add
